New API and Flatpak backend code
Design a new and better API and adapt the Flatpak backend
immediately.

Move the backend code into plugins. At the moment we only deal
with Flatpak because that's what we are concerned with
at Liri.

Add more features to the Flatpak backend such as full metadata
download and parsing, transactional operations like install,
uninstall and updates.

Issue: #6