Transaction overhaul
You can now decide whether a transaction can be cancelled when you create it,
this works only if the backend can cancel transactions of course.

Transactions now emit a specific signal when they succeed or fail and
we route it to the app signal for installation, uninstallation or update
depending on the transaction type.

Previously we emitted such signals right after running the transaction,
without waiting for its execution.

Installation, uninstallation and update methods of a resource now return
the transaction and that's used to create the UI.