Improve sources API
Allow backends to redefine getter and setter methods and eliminate internal
book keeping.

This also incidentally fixes disabling Flatpak sources.
With Flatpak we disable a source but our API is designed to enable
a source, hence the flag wasn't read correcly.

Specify the backend at source creation time and don't allow to
be changed later on.

Add the following new properties to sources:

 * comment
 * description
 * gpgVerify
 * iconUrl