flatpak: Fix the number of updates
Count only the resources that we are going to actually set
as upgradable.

Emit the signal everytime, even though the number of resources to update
is not changed.  After all we are emitting the signal in response
to a check for updates, this means we want to know if there are upgradable
resources every single time.