Reviews and ratings
We add a new type of backend, called ReviewsBackend that allows us to
write plugins to handle ratings and reviews.

This means that the previous plugins are now references as `resources`
backends, while the new one are `reviews` backends.

Plugins are installed in different paths.  At startup, the resources
backends are initialized before the reviews backends.

There is a ReviewsModel that is populated asynchronously with reviews
of a specific software resource.  This model is used to display
the reviews of an app.

We implement ODRS for ratings and reviews.

Closes: #13