Consolidate resources into one entry
We might have the same resource from many different sources, but we want
to present it to the user only once.

Associate each resource with their source.

Create a ResourceProxy that contains all the resources and lets you select
the resource to show corresponding to the selected source.

Proxies are stored inside SoftwareManager and emits signals that are used
by ResourcesModel to insert or remove rows.

Resources are now added to SoftwareManager all at once when they are already
refined, this is also faster than adding one resource at a time.

Closes: #20